Started to reorganize project structure (broke something along the way)
