+++ src/CMakeLists.txt
@@ -35,7 +35,7 @@ set(MAPHOON_IDEEX ${SOURCE_DIR}/idee.x)
 set(MAPHOONGEN token.cpp parser.cpp)
 prepend(MAPHOONGEN ${MAPHOON_WORKDIR}/ ${MAPHOONGEN})
 
-set(MAPHOON_INPUT ${CMAKE_CURRENT_SOURCE_DIR}/grammar.m)
+set(MAPHOON_INPUT ${CMAKE_CURRENT_SOURCE_DIR}/parsing/grammar.m)
 
 # Use maphoon to generate the parser
 add_custom_command(
@@ -48,9 +48,9 @@ add_custom_command(
   COMMENT "Generating parser with Maphoon"
   )
 
-FLEX_TARGET(lexer lexer.lex ${CMAKE_CURRENT_BINARY_DIR}/lexer.cpp)
+FLEX_TARGET(lexer parsing/lexer.lex ${CMAKE_CURRENT_BINARY_DIR}/lexer.cpp)
 
-file(GLOB SOURCES
+file(GLOB_RECURSE SOURCES
   ./*.cpp
   )
 
@@ -61,6 +61,11 @@ set(LLVM_LIBRARIES LLVMCore LLVMScalarOpts)
 # Adding sources dir to include list, so that files generated by maphoon can see our headers and vice versa
 include_directories(
   ${CMAKE_CURRENT_SOURCE_DIR}
+  ${CMAKE_CURRENT_SOURCE_DIR}/parsing
+  ${CMAKE_CURRENT_SOURCE_DIR}/ast
+  ${CMAKE_CURRENT_SOURCE_DIR}/misc
+  ${CMAKE_CURRENT_SOURCE_DIR}types
+  ${CMAKE_CURRENT_SOURCE_DIR}/world
   ${CMAKE_CURRENT_BINARY_DIR}
   ${LLVM_INCLUDE_DIRS}
   )
@@ -102,4 +107,4 @@ target_link_libraries(
 add_custom_command(
   TARGET ${CMAKE_PROJECT_NAME} POST_BUILD
   COMMAND ${CMAKE_COMMAND} -E copy "${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/${CMAKE_PROJECT_NAME}${CMAKE_EXECUTABLE_SUFFIX}" "${CMAKE_BINARY_DIR}/"
-  )
+)
