+++ src/compiler.cpp
@@ -1,9 +1,9 @@
 #include "compiler.h"
-#include "utils.h"
-#include "codegencontext.h"
-#include "tree.h"
+#include "misc/utils.h"
+#include "world/codegencontext.h"
+#include "ast/tree.h"
 #include "token.h"
-#include "tokenizer.h"
+#include "parsing/tokenizer.h"
 #include "parser.h"
 
 #include "llvm/IR/Module.h"
