+++ src/misc/cast.h
