+++ src/parsing/grammar.m
@@ -139,17 +139,17 @@
 
 
 %intokenheader #include <memory>
-%intokenheader #include "tree.h"
-%intokenheader #include "types.h"
+%intokenheader #include "../ast/tree.h"
+%intokenheader #include "../types/types.h"
 
 %global ctx ccoscope::CodegenContext
 
 #include <cassert>
 #define ASSERT( X ) { assert( ( X ) ); }
 #include <cstdio>
-#include "tree.h"
-#include "types.h"
-#include "codegencontext.h"
+#include "../ast/tree.h"
+#include "../types/types.h"
+#include "../world/codegencontext.h"
 #include <memory>
 
 % Start : Module
