+++ src/types/conversion.h
@@ -1,7 +1,7 @@
-#ifndef __CONVERSIONS_H__
-#define __CONVERSIONS_H__
+#ifndef __CONVERSION_H__
+#define __CONVERSION_H__
 
-#include "proxy.h"
+#include "../misc/proxy.h"
 
 #include "llvm/IR/Value.h"
 
@@ -25,4 +25,4 @@ struct Conversion{
 
 }
 
-#endif // __CONVERSIONS_H__
+#endif // __CONVERSION_H__
