+++ src/types/typematcher.cpp
@@ -1,5 +1,5 @@
 #include "typematcher.h"
-#include "codegencontext.h"
+#include "../world/codegencontext.h"
 #include <algorithm>
 #include <queue>
