+++ src/types/typematcher.h
@@ -2,7 +2,7 @@
 #ifndef __TYPEMATCHER_H__
 #define __TYPEMATCHER_H__
 
-#include "conversions.h"
+#include "conversion.h"
 #include "types.h"
 #include <string>
 #include <list>
